refactor: 시간 생성, 삭제를 관리자만 할 수 있도록 변경

+++ src/main/java/roomescape/presentation/rest/TimeSlotController.java
@@ -17,6 +17,8 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.application.TimeSlotService;
+import roomescape.domain.auth.AuthenticationInfo;
+import roomescape.exception.AuthorizationException;
 import roomescape.presentation.request.CreateTimeSlotRequest;
 import roomescape.presentation.response.AvailableTimeSlotResponse;
 import roomescape.presentation.response.TimeSlotResponse;
@@ -30,7 +32,13 @@ public class TimeSlotController {
 
     @PostMapping
     @ResponseStatus(CREATED)
-    public TimeSlotResponse register(@RequestBody @Valid final CreateTimeSlotRequest request) {
+    public TimeSlotResponse register(
+        final AuthenticationInfo authenticationInfo,
+        @RequestBody @Valid final CreateTimeSlotRequest request
+    ) {
+        if (authenticationInfo.isNotAdmin()) {
+            throw new AuthorizationException("관리자에게만 허용된 작업입니다.");
+        }
         var timeSlot = service.register(request.startAt());
         return TimeSlotResponse.from(timeSlot);
     }
@@ -43,8 +51,8 @@ public class TimeSlotController {
 
     @GetMapping(value = "/available", params = {"date", "themeId"})
     public List<AvailableTimeSlotResponse> getAvailableTimes(
-            @RequestParam("date") final LocalDate date,
-            @RequestParam("themeId") final Long themeId
+        @RequestParam("date") final LocalDate date,
+        @RequestParam("themeId") final Long themeId
     ) {
         var availableTimeSlots = service.findAvailableTimeSlots(date, themeId);
         return AvailableTimeSlotResponse.from(availableTimeSlots);
@@ -52,7 +60,13 @@ public class TimeSlotController {
 
     @DeleteMapping("/{id}")
     @ResponseStatus(NO_CONTENT)
-    public void delete(@PathVariable("id") final long id) {
+    public void delete(
+        final AuthenticationInfo authenticationInfo,
+        @PathVariable("id") final long id
+    ) {
+        if (authenticationInfo.isNotAdmin()) {
+            throw new AuthorizationException("관리자에게만 허용된 작업입니다.");
+        }
         service.removeById(id);
     }
 }
